Trait pallet_author_mapping::pallet::Config

pub trait Config: Config {
    type Event: From<Event<Self>> + IsType<Self::Event>;
    type AuthorId: Member + Parameter + MaybeSerializeDeserialize;
    type DepositCurrency: Currency<Self::AccountId> + ReservableCurrency<Self::AccountId>;
    type DepositAmount: Get<<Self::DepositCurrency as Currency<Self::AccountId>>::Balance>;
    fn can_register(account: &Self::AccountId) -> bool;
}

Configuration trait of this pallet. We tightly couple to Parachain Staking in order to ensure that only staked accounts can create registrations in the first place. This could be generalized.

Associated Types

type Event: From<Event<Self>> + IsType<Self::Event>

Overarching event type

type AuthorId: Member + Parameter + MaybeSerializeDeserialize

The type of authority id that will be used at the conensus layer.

type DepositCurrency: Currency<Self::AccountId> + ReservableCurrency<Self::AccountId>

Currency in which the security deposit will be taken.

type DepositAmount: Get<<Self::DepositCurrency as Currency<Self::AccountId>>::Balance>

The amount that should be taken as a security deposit when registering an AuthorId.

Loading content...
Required methods

fn can_register(account: &Self::AccountId) -> bool

A rough preliminary check to determine whether an account can make a new registration. If you don’t wish to do any such check, just return true.

Loading content...