Trait pallet_author_mapping::pallet::Config

pub trait Config: Config {
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type DepositCurrency: Currency<Self::AccountId> + ReservableCurrency<Self::AccountId>;
    type DepositAmount: Get<<Self::DepositCurrency as Currency<Self::AccountId>>::Balance>;
    type Keys: Parameter + Member + MaybeSerializeDeserialize + From<NimbusId>;
    type WeightInfo: WeightInfo;
}

Configuration trait of this pallet

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

Overarching event type

type DepositCurrency: Currency<Self::AccountId> + ReservableCurrency<Self::AccountId>

Currency in which the security deposit will be taken.

type DepositAmount: Get<<Self::DepositCurrency as Currency<Self::AccountId>>::Balance>

The amount that should be taken as a security deposit when registering a NimbusId.

type Keys: Parameter + Member + MaybeSerializeDeserialize + From<NimbusId>

Additional keys Convertible From to get default keys for each mapping (for the migration)

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.