Trait pallet_author_mapping::pallet::Config

pub trait Config: Config {
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type DepositCurrency: Currency<Self::AccountId> + ReservableCurrency<Self::AccountId>;
    type DepositAmount: Get<<Self::DepositCurrency as Currency<Self::AccountId>>::Balance>;
    type WeightInfo: WeightInfo;
}

Configuration trait of this pallet. We tightly couple to Parachain Staking in order to ensure that only staked accounts can create registrations in the first place. This could be generalized.

Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

Overarching event type

type DepositCurrency: Currency<Self::AccountId> + ReservableCurrency<Self::AccountId>

Currency in which the security deposit will be taken.

type DepositAmount: Get<<Self::DepositCurrency as Currency<Self::AccountId>>::Balance>

The amount that should be taken as a security deposit when registering a NimbusId.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.